Revert "[BDGR-180][WIP] Write new vMix client and mock server (#645)"
This reverts commit d2b8bf73dd1cd5b5dfc6dd690b6dfef2363f01f5.
