+++ desktop/e2e/complete/vmix.spec.ts
@@ -7,7 +7,7 @@ import {
   server,
 } from "./serverAPI";
 import type { CompleteShowType } from "../../src/common/types";
-import type VMixConnection from "@ystv/vmix";
+import type VMixConnection from "../../src/main/vmix/vmix";
 import { PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
 
 let testShow: CompleteShowType;
