+++ desktop/package.json
@@ -45,7 +45,6 @@
     "@types/react-dom": "^18.2.14",
     "@types/uuid": "^9.0.2",
     "@types/which": "^3.0.0",
-    "@ystv/vmix": "workspace:*",
     "badger-server": "workspace:*",
     "bufferutil": "^4.0.7",
     "classnames": "^2.3.2",
