+++ desktop/src/main/vmix/vMix.mock.ts
@@ -1,6 +1,6 @@
 import { getLogger } from "loglevel";
 import { mock, when, It, verify, reset } from "strong-mock";
-import VMixConnection from "@ystv/vmix";
+import VMixConnection from "./vmix";
 
 const logger = getLogger("vMix.mock");
