--- utility/vmix/README.md
@@ -1,1 +1,0 @@
-# vmix
