--- utility/vmix/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@ystv/vmix",
-  "description": "Strongly typed vMix API client and testing utilities",
-  "author": "Marks Polakovs <marks.polakovs@ystv.co.uk>",
-  "packageManager": "yarn@4.3.1",
-  "exports": {
-    "import": {
-      "badger-internal": "./src/index.ts",
-      "default": "./dist/index.js",
-      "types": "./dist/index.d.ts"
-    },
-    "require": {
-      "badger-internal": "./src/index.js",
-      "default": "./dist/index.js",
-      "types": "./dist/index.d.ts"
-    }
-  },
-  "scripts": {
-    "test": "jest",
-    "typecheck": "tsc --noEmit",
-    "build": "rollup -c rollup.config.mjs"
-  },
-  "dependencies": {
-    "qs": "^6.13.0",
-    "uuid": "^10.0.0",
-    "xml-js": "^1.6.11",
-    "zod": "^3.23.8"
-  },
-  "devDependencies": {
-    "@jest/globals": "^29.7.0",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "@rollup/plugin-typescript": "^11.1.6",
-    "@tsconfig/node-lts": "^20.1.3",
-    "@types/jest": "^29.5.12",
-    "@types/node": "^22.1.0",
-    "@xmldom/xmldom": "^0.8.10",
-    "esbuild": "^0.23.1",
-    "jest": "^29.7.0",
-    "rollup": "^4.21.0",
-    "rollup-plugin-esbuild": "^6.1.1",
-    "rollup-plugin-node-externals": "^7.1.3",
-    "ts-jest": "^29.2.4",
-    "typescript": "^5.5.4",
-    "xpath": "^0.0.34"
-  },
-  "peerDependencies": {
-    "immer": "*"
-  },
-  "peerDependenciesMeta": {
-    "immer": {
-      "optional": true
-    }
-  }
-}
