--- utility/vmix/src/index.ts
@@ -1,8 +1,0 @@
-import VMixConnection from "./vmix";
-
-import { VMixState } from "./vmixState";
-
-export * from "./vmixTypes";
-
-export default VMixConnection;
-export { VMixConnection, VMixState };
