--- utility/vmix/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "@tsconfig/node-lts/tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "declaration": true,
-    "declarationMap": true,
-    "composite": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["src/**/*.test.ts"]
-}
